import.py: Raise an error when no files are found
Previously the error looked like this:

   An unhandled exception occured:

    Traceback (most recent call last):
      File "/usr/lib64/python3.5/shutil.py", line 544, in move
        os.rename(src, real_dst)
    FileNotFoundError: [Errno 2] No such file or directory: '/home/shared/cache/buildstream/build/gnu-toolchain-base-sdk-c8dedylv/input/files' -> '/home/shared/cache/buildstream/build/gnu-toolchain-base-sdk-c8dedylv/output/usr'

    During handling of the above exception, another exception occurred:

    Traceback (most recent call last):
      File "/home/shared/src/buildstream/buildstream/_scheduler/job.py", line 255, in child_action
        result = self.action(element)
      File "/home/shared/src/buildstream/buildstream/_scheduler/buildqueue.py", line 34, in process
        element._assemble()
      File "/home/shared/src/buildstream/buildstream/element.py", line 718, in _assemble
        collect = self.assemble(sandbox)
      File "/home/shared/src/buildstream/buildstream/plugins/elements/import.py", line 75, in assemble
        shutil.move(inputdir, outputdir)
      File "/usr/lib64/python3.5/shutil.py", line 558, in move
        copy_function(src, real_dst)
      File "/usr/lib64/python3.5/shutil.py", line 257, in copy2
        copyfile(src, dst, follow_symlinks=follow_symlinks)
      File "/usr/lib64/python3.5/shutil.py", line 120, in copyfile
        with open(src, 'rb') as fsrc:
    FileNotFoundError: [Errno 2] No such file or directory: '/home/shared/cache/buildstream/build/gnu-toolchain-base-sdk-c8dedylv/input/files'

Now it looks like this:

    import element at gnu-toolchain/base-sdk.bst [line 1 column 0]: No files were found at path 'files'

    Sandbox directory: /home/shared/cache/buildstream/build/gnu-toolchain-base-sdk-3dy3eozv

+++ buildstream/plugins/elements/import.py
@@ -32,7 +32,7 @@ The empty configuration is as such:
 
 import os
 import shutil
-from buildstream import BuildElement
+from buildstream import BuildElement, ElementError
 
 
 # Element implementation for the 'import' kind.
@@ -76,6 +76,10 @@ class ImportElement(BuildElement):
 
         # Ensure target directory parent
         os.makedirs(os.path.dirname(outputdir), exist_ok=True)
+
+        if not os.path.exists(inputdir):
+            raise ElementError("{}: No files were found at path '{}'"
+                               .format(self, self.source))
 
         # Move it over
         shutil.move(inputdir, outputdir)
